pyproject.toml
[project]
name = "still-to-life-background"
version = "0.1.0"
description = "Turn images into moving macOS wallpapers using AI depth parallax"
requires-python = ">=3.12,<3.14"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "spandrel>=0.4",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "packaging>=23.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "pillow>=10.0",
]
